pyproject.toml
[tool.poetry]
name = "retfound-mae"
version = "0.1.0"
description = ""
authors = ["Mason Cleveland <mcleveland2@mgh.harvard.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
# python = "^3.11"
# torch = "^2.3.1"
# torchvision = "^0.18.1"
# torchaudio = "^2.3.1"
# opencv-python = "^4.10.0.84"
# pandas = "^2.2.2"
# pillow = "^10.3.0"
# protobuf = "<3.19.0"
# pycm = "^4.0"
# pydicom = "^2.4.4"
# scikit-image = "^0.24.0"
# scikit-learn = "^1.5.0"
# scipy = "^1.13.1"
# tensorboard = "<2.17.0"
# tensorboard-data-server = "^0.7.2"
# tensorboard-plugin-wit = "^1.8.1"
# timm = "^1.0.3"
# tqdm = "^4.66.4"
# visidata = "^3.0.2"
# numpy = "<1.20.0"

python = ">=3.10"
torch = "<2.0.0"
torchvision = ">0.9.1"
torchaudio = ">0.8.1"
opencv-python = ">4.5.3.56"
pandas = ">0.25.3"
Pillow = ">8.3.1"
protobuf = ">3.17.3"
pycm = ">3.2"
pydicom = ">2.3.0"
scikit-image = ">0.17.2"
scikit-learn = ">0.24.2"
scipy = ">1.5.4"
numpy = "<2.0.0"
tensorboard = "*"
tensorboard-data-server = "*"
tensorboard-plugin-wit = "*"
timm = "*"
tqdm = ">4.62.1"
matplotlib = "*"
visidata = "*"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
